plugin: Also coerce milliseconds to float, not int

+++ extras/supybot/GNOMEOSTree/plugin.py
@@ -91,7 +91,7 @@ class GNOMEOSTree(callbacks.Plugin):
             msg = "Current " + taskname
         success = metadata['success']
         success_str = success and 'successful' or 'failed'
-        millis = int(metadata['elapsedMillis'])
+        millis = float(metadata['elapsedMillis'])
         msg += " %s: %s in %.1f seconds. " % (taskver, success_str, millis / 1000.0)
 
         status_path = os.path.join(current_task_path, 'status.txt')
